Update public-api.md
Not sure about the proper externalization of the shutdown scenario. if we want to hide the fact that a node is being recycled we may want to return the currently known status of each transaction which will lead to the client retrying the call. a different approach is to return a "server shutting down" error which should also result in a retry from clients.